fix: Unify stakable token filtering to prevent empty list
