+++ apps/extension/src/pages/stake/empty.tsx
@@ -6,9 +6,7 @@ import { ColorPalette } from "../../styles";
 import { Box } from "../../components/box";
 import { Gutter } from "../../components/gutter";
 import { Subtitle2 } from "../../components/typography";
-import { useStore } from "../../stores";
 import { TokenItem } from "../main/components";
-import { Dec } from "@keplr-wallet/unit";
 import { TextButton } from "../../components/button-text";
 import { useNavigate } from "react-router";
 import { ChevronRightIcon } from ".";
@@ -16,44 +14,13 @@ import { FormattedMessage, useIntl } from "react-intl";
 import { MainH1 } from "../../components/typography/main-h1";
 import { useGetStakingApr } from "../../hooks/use-get-staking-apr";
 import { EarnRewardsIcon } from "./components/earn-rewards-icon";
-import { useKcrStakingUrls } from "../../hooks/use-kcr-staking-urls";
-
-const zeroDec = new Dec(0);
+import { useStakableTokens } from "./hooks/use-stakable-tokens";
 
 export const StakeEmptyPage: FunctionComponent = observer(() => {
   const theme = useTheme();
   const navigate = useNavigate();
   const intl = useIntl();
-  const { chainStore, hugeQueriesStore, priceStore } = useStore();
-  const { getKcrStakingUrl, hasKcrStakingUrl } = useKcrStakingUrls();
-
-  const stakableTokens = hugeQueriesStore.stakables
-    .filter((token) => {
-      if (!token.token.toDec().gt(zeroDec)) {
-        return false;
-      }
-      if ("starknet" in token.chainInfo) {
-        return true;
-      }
-      if ("bitcoin" in token.chainInfo) {
-        return false;
-      }
-      const chainInfo = chainStore.getChain(token.chainInfo.chainId);
-      const hasNativeUrl =
-        !!chainInfo.embedded.embedded &&
-        !!chainInfo.embedded.walletUrlForStaking;
-      return hasNativeUrl || hasKcrStakingUrl(token.chainInfo.chainId);
-    })
-    .sort((a, b) => {
-      const aPrice = priceStore.calculatePrice(a.token)?.toDec() ?? zeroDec;
-      const bPrice = priceStore.calculatePrice(b.token)?.toDec() ?? zeroDec;
-
-      if (aPrice.equals(bPrice)) {
-        return 0;
-      }
-      return aPrice.gt(bPrice) ? -1 : 1;
-    })
-    .slice(0, 4);
+  const { stakableTokens, getStakingUrl } = useStakableTokens();
 
   return (
     <MainHeaderLayout>
@@ -98,17 +65,7 @@ export const StakeEmptyPage: FunctionComponent = observer(() => {
         <Gutter size="1rem" />
 
         {stakableTokens.map((viewToken) => {
-          const stakingUrl = (() => {
-            if ("starknet" in viewToken.chainInfo) {
-              return "https://dashboard.endur.fi/stake";
-            }
-            const chainInfo = chainStore.getChain(viewToken.chainInfo.chainId);
-            return (
-              chainInfo.embedded.walletUrlForStaking ||
-              getKcrStakingUrl(viewToken.chainInfo.chainId)
-            );
-          })();
-
+          const stakingUrl = getStakingUrl(viewToken);
           const stakingAprDec = useGetStakingApr(viewToken.chainInfo.chainId);
 
           return (
