+++ apps/extension/src/pages/stake/hooks/use-kcr-staking-urls.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { useStore } from "../stores";
+import { useStore } from "../../../stores";
 import { ChainInfo } from "@keplr-wallet/types";
 import { ChainIdHelper } from "@keplr-wallet/cosmos";
